refactor stencil state management code
