+++ cocos/2d/CCClippingNode.h
@@ -34,6 +34,8 @@
 #include "renderer/CCCustomCommand.h"
 
 NS_CC_BEGIN
+
+class StencilStateManager;
 /**
  *  @addtogroup _2d
  *  @{
@@ -153,34 +155,9 @@ CC_CONSTRUCTOR_ACCESS:
     virtual bool init(Node *stencil);
 
 protected:
-    /**draw fullscreen quad to clear stencil bits
-    */
-    void drawFullScreenQuadClearStencil();
-
     Node* _stencil;
-    GLfloat _alphaThreshold;
-    bool    _inverted;
-
-    //renderData and callback
-    void onBeforeVisit();
-    void onAfterDrawStencil();
-    void onAfterVisit();
-
-    GLboolean _currentStencilEnabled;
-    GLuint _currentStencilWriteMask;
-    GLenum _currentStencilFunc;
-    GLint _currentStencilRef;
-    GLuint _currentStencilValueMask;
-    GLenum _currentStencilFail;
-    GLenum _currentStencilPassDepthFail;
-    GLenum _currentStencilPassDepthPass;
-    GLboolean _currentDepthWriteMask;
-
-    GLboolean _currentAlphaTestEnabled;
-    GLenum _currentAlphaTestFunc;
-    GLclampf _currentAlphaTestRef;
-
-    GLint _mask_layer_le;
+   
+    StencilStateManager* _stencilStateManager;
     
     GroupCommand _groupCommand;
     CustomCommand _beforeVisitCmd;
