+++ cocos/ui/UILayout.cpp
@@ -74,18 +74,7 @@ _clippingStencil(nullptr),
 _clippingRect(Rect::ZERO),
 _clippingParent(nullptr),
 _clippingRectDirty(true),
-_currentStencilEnabled(GL_FALSE),
-_currentStencilWriteMask(~0),
-_currentStencilFunc(GL_ALWAYS),
-_currentStencilRef(0),
-_currentStencilValueMask(~0),
-_currentStencilFail(GL_KEEP),
-_currentStencilPassDepthFail(GL_KEEP),
-_currentStencilPassDepthPass(GL_KEEP),
-_currentDepthWriteMask(GL_TRUE),
-_currentAlphaTestEnabled(GL_FALSE),
-_currentAlphaTestFunc(GL_ALWAYS),
-_currentAlphaTestRef(1),
+_stencileStateManager(new StencilStateManager()),
 _doLayoutDirty(true),
 _isInterceptTouch(false),
 _loopFocus(false),
@@ -264,13 +253,13 @@ void Layout::stencilClippingVisit(Renderer *renderer, const Mat4& parentTransfor
     renderer->pushGroup(_groupCommand.getRenderQueueID());
     
     _beforeVisitCmdStencil.init(_globalZOrder);
-    _beforeVisitCmdStencil.func = CC_CALLBACK_0(Layout::onBeforeVisitStencil, this);
+    _beforeVisitCmdStencil.func = CC_CALLBACK_0(StencilStateManager::onBeforeVisit, _stencileStateManager);
     renderer->addCommand(&_beforeVisitCmdStencil);
     
     _clippingStencil->visit(renderer, _modelViewTransform, flags);
     
     _afterDrawStencilCmd.init(_globalZOrder);
-    _afterDrawStencilCmd.func = CC_CALLBACK_0(Layout::onAfterDrawStencil, this);
+    _afterDrawStencilCmd.func = CC_CALLBACK_0(StencilStateManager::onAfterDrawStencil, _stencileStateManager);
     renderer->addCommand(&_afterDrawStencilCmd);
     
     int i = 0;      // used by _children
@@ -318,146 +307,12 @@ void Layout::stencilClippingVisit(Renderer *renderer, const Mat4& parentTransfor
 
     
     _afterVisitCmdStencil.init(_globalZOrder);
-    _afterVisitCmdStencil.func = CC_CALLBACK_0(Layout::onAfterVisitStencil, this);
+    _afterVisitCmdStencil.func = CC_CALLBACK_0(StencilStateManager::onAfterVisit, _stencileStateManager);
     renderer->addCommand(&_afterVisitCmdStencil);
     
     renderer->popGroup();
     
     director->popMatrix(MATRIX_STACK_TYPE::MATRIX_STACK_MODELVIEW);
-}
-    
-void Layout::onBeforeVisitStencil()
-{
-    auto s_layer = StencilBitsManager::getInstance()->getStencilLayerMask();
-    s_layer++;
-    StencilBitsManager::getInstance()->setStencilLayerMask(s_layer);
-    GLint mask_layer = 0x1 << s_layer;
-    GLint mask_layer_l = mask_layer - 1;
-    _mask_layer_le = mask_layer | mask_layer_l;
-    _currentStencilEnabled = glIsEnabled(GL_STENCIL_TEST);
-    glGetIntegerv(GL_STENCIL_WRITEMASK, (GLint *)&_currentStencilWriteMask);
-    glGetIntegerv(GL_STENCIL_FUNC, (GLint *)&_currentStencilFunc);
-    glGetIntegerv(GL_STENCIL_REF, &_currentStencilRef);
-    glGetIntegerv(GL_STENCIL_VALUE_MASK, (GLint *)&_currentStencilValueMask);
-    glGetIntegerv(GL_STENCIL_FAIL, (GLint *)&_currentStencilFail);
-    glGetIntegerv(GL_STENCIL_PASS_DEPTH_FAIL, (GLint *)&_currentStencilPassDepthFail);
-    glGetIntegerv(GL_STENCIL_PASS_DEPTH_PASS, (GLint *)&_currentStencilPassDepthPass);
-    
-    glEnable(GL_STENCIL_TEST);
-//    RenderState::StateBlock::_defaultState->setStencilTest(true);
-
-    CHECK_GL_ERROR_DEBUG();
-    glStencilMask(mask_layer);
-//    RenderState::StateBlock::_defaultState->setStencilWrite(mask_layer);
-
-    glGetBooleanv(GL_DEPTH_WRITEMASK, &_currentDepthWriteMask);
-
-    glDepthMask(GL_FALSE);
-    RenderState::StateBlock::_defaultState->setDepthWrite(false);
-
-    glStencilFunc(GL_NEVER, mask_layer, mask_layer);
-    glStencilOp(GL_ZERO, GL_KEEP, GL_KEEP);
-
-
-    this->drawFullScreenQuadClearStencil();
-    
-    glStencilFunc(GL_NEVER, mask_layer, mask_layer);
-//    RenderState::StateBlock::_defaultState->setStencilFunction(
-//                                                               RenderState::STENCIL_NEVER,
-//                                                               mask_layer,
-//                                                               mask_layer);
-
-    glStencilOp(GL_REPLACE, GL_KEEP, GL_KEEP);
-//    RenderState::StateBlock::_defaultState->setStencilOperation(
-//                                                                RenderState::STENCIL_OP_REPLACE,
-//                                                                RenderState::STENCIL_OP_KEEP,
-//                                                                RenderState::STENCIL_OP_KEEP);
-}
-    
-void Layout::drawFullScreenQuadClearStencil()
-{
-    Director* director = Director::getInstance();
-    CCASSERT(nullptr != director, "Director is null when setting matrix stack");
-
-    director->pushMatrix(MATRIX_STACK_TYPE::MATRIX_STACK_MODELVIEW);
-    director->loadIdentityMatrix(MATRIX_STACK_TYPE::MATRIX_STACK_MODELVIEW);
-
-    director->pushMatrix(MATRIX_STACK_TYPE::MATRIX_STACK_PROJECTION);
-    director->loadIdentityMatrix(MATRIX_STACK_TYPE::MATRIX_STACK_PROJECTION);
-    
-    Vec2 vertices[] =
-    {
-        Vec2(-1, -1),
-        Vec2(1, -1),
-        Vec2(1, 1),
-        Vec2(-1, 1)
-    };
-    
-    auto glProgram = GLProgramCache::getInstance()->getGLProgram(GLProgram::SHADER_NAME_POSITION_U_COLOR);
-    
-    int colorLocation = glProgram->getUniformLocation("u_color");
-    CHECK_GL_ERROR_DEBUG();
-    
-    Color4F color(1, 1, 1, 1);
-    
-    glProgram->use();
-    glProgram->setUniformsForBuiltins();
-    glProgram->setUniformLocationWith4fv(colorLocation, (GLfloat*) &color.r, 1);
-    
-    GL::enableVertexAttribs( GL::VERTEX_ATTRIB_FLAG_POSITION );
-    
-    glVertexAttribPointer(GLProgram::VERTEX_ATTRIB_POSITION, 2, GL_FLOAT, GL_FALSE, 0, vertices);
-    glDrawArrays(GL_TRIANGLE_FAN, 0, 4);
-    
-    CC_INCREMENT_GL_DRAWN_BATCHES_AND_VERTICES(1, 4);
-    
-    director->popMatrix(MATRIX_STACK_TYPE::MATRIX_STACK_PROJECTION);
-    director->popMatrix(MATRIX_STACK_TYPE::MATRIX_STACK_MODELVIEW);
-}
-
-void Layout::onAfterDrawStencil()
-{
-    glDepthMask(_currentDepthWriteMask);
-    RenderState::StateBlock::_defaultState->setDepthWrite(_currentDepthWriteMask != 0);
-
-    glStencilFunc(GL_EQUAL, _mask_layer_le, _mask_layer_le);
-//    RenderState::StateBlock::_defaultState->setStencilFunction(
-//                                                                RenderState::STENCIL_EQUAL,
-//                                                               _mask_layer_le,
-//                                                               _mask_layer_le);
-
-    glStencilOp(GL_KEEP, GL_KEEP, GL_KEEP);
-//    RenderState::StateBlock::_defaultState->setStencilOperation(
-//                                                                RenderState::STENCIL_OP_KEEP,
-//                                                                RenderState::STENCIL_OP_KEEP,
-//                                                                RenderState::STENCIL_OP_KEEP);
-
-}
-
-
-void Layout::onAfterVisitStencil()
-{
-    glStencilFunc(_currentStencilFunc, _currentStencilRef, _currentStencilValueMask);
-//    RenderState::StateBlock::_defaultState->setStencilFunction(
-//                                                               (RenderState::StencilFunction)_currentStencilFunc,
-//                                                               _currentStencilRef,
-//                                                               _currentStencilValueMask);
-
-    glStencilOp(_currentStencilFail, _currentStencilPassDepthFail, _currentStencilPassDepthPass);
-//    RenderState::StateBlock::_defaultState->setStencilOperation(
-//                                                                (RenderState::StencilOperation)_currentStencilFail,
-//                                                                (RenderState::StencilOperation)_currentStencilPassDepthFail,
-//                                                                (RenderState::StencilOperation)_currentStencilPassDepthPass);
-
-    glStencilMask(_currentStencilWriteMask);
-    if (!_currentStencilEnabled)
-    {
-        glDisable(GL_STENCIL_TEST);
-//        RenderState::StateBlock::_defaultState->setStencilTest(false);
-    }
-    auto s_layer = StencilBitsManager::getInstance()->getStencilLayerMask();
-    s_layer--;
-    StencilBitsManager::getInstance()->setStencilLayerMask(s_layer);
 }
     
 void Layout::onBeforeVisitScissor()
@@ -532,16 +387,6 @@ void Layout::setClippingEnabled(bool able)
         case ClippingType::STENCIL:
             if (able)
             {
-                static bool once = true;
-                if (once)
-                {
-                    glGetIntegerv(GL_STENCIL_BITS, &g_sStencilBits);
-                    if (g_sStencilBits <= 0)
-                    {
-                        CCLOG("Stencil buffer is not enabled.");
-                    }
-                    once = false;
-                }
                 _clippingStencil = DrawNode::create();
                 if (_running)
                 {
